+++ libraries/provider_container.rb
@@ -25,7 +25,12 @@ class Chef
         @current_resource
       end
 
-      ## add extra build parameters and create new container ##
+
+
+
+      ##
+      ## add extra build parameters and create new container
+      ##
 
       def create_unique_container
         @container_create_options << %Q{--hostname="#{new_resource.service_name}"}
@@ -36,7 +41,9 @@ class Chef
         return DockerWrapper::Container.create((@container_create_options + new_resource.container_create_options).join(' '), "#{new_resource.base_image}:#{new_resource.base_image_tag}")
       end
 
-      ## create chef credentials and start container ##
+      ##
+      ## create chef credentials and start container
+      ##
 
       def start_container(container)
         populate_chef_secure_path
@@ -44,7 +51,9 @@ class Chef
         container.start
       end
 
-      ## stop container, try killing ##
+      ##
+      ## stop container, try killing
+      ##
 
       def stop_container(container)
         Chef::Log.info("Stopping container #{container.id}...")
@@ -53,7 +62,9 @@ class Chef
         raise StopContainer, "Unable to stop container #{container.name}" if container.running?
       end
 
-      ## stop and remove container ##
+      ##
+      ## stop and remove container
+      ##
 
       def remove_container(container)
         image = DockerWrapper::Image.new(container.parent_id)
@@ -68,7 +79,9 @@ class Chef
         end
       end
 
-      ## path for writing CID and other files ##
+      ##
+      ## path for writing CID and other files
+      ##
 
       def cache_path
         return @cache_resources unless @cache_resources.nil?
@@ -79,7 +92,9 @@ class Chef
         return @cache_resources
       end
 
-      ## path for chef keys. mounted to container ##
+      ##
+      ## path for chef keys. mounted to container
+      ##
 
       def chef_secure_path
         return @secure_resources unless @secure_resources.nil?
@@ -90,7 +105,9 @@ class Chef
         return @secure_resources
       end
 
-      ## chef client key file ##
+      ##
+      ## chef client key file
+      ##
 
       def client_key_file
         return @client_key_file unless @client_key_file.nil?
@@ -99,7 +116,9 @@ class Chef
         return @client_key_file
       end
 
-      ## write CID file to cache path ##
+      ##
+      ## write CID file to cache path
+      ##
 
       def populate_cache_path(container)
         cache_path.run_action(:create)
@@ -109,7 +128,9 @@ class Chef
         r.run_action(:create)
       end
 
-      ## write chef keys to secure path ##
+      ##
+      ## write chef keys to secure path
+      ##
 
       def populate_chef_secure_path
         chef_secure_path.run_action(:create)
@@ -137,33 +158,43 @@ class Chef
         end
       end
 
-      ## remove cache path and CID file ##
+      ##
+      ## remove cache path and CID file
+      ##
 
       def remove_cache_path
         cache_path.run_action(:delete)
       end
 
-      ## remove chef node for this container ##
+      ##
+      ## remove chef node for this container
+      ##
 
       def remove_chef_node
         remove_from_chef(new_resource.chef_server_url, new_resource.service_name, client_key_file)
       end
 
-      ## remove chef keys and secure path ##
+      ##
+      ## remove chef keys and secure path
+      ##
 
       def remove_chef_secure_path
         remove_chef_node
         chef_secure_path.run_action(:delete)
       end
 
-      ## write container name and ID mapping to service name ##
+      #
+      ## write container name and ID mapping to service name
+      ##
 
       def set_service_mapping(container)
         node.default['docker_deploy']['service_mapping'][new_resource.service_name]['id'] = container.id
         node.default['docker_deploy']['service_mapping'][new_resource.service_name]['name'] = container.name
       end
 
-      ## remove container name from config so that they can be compared ##
+      ##
+      ## remove container name from config so that they can be compared
+      ##
 
       def clean_hostconfig(container)
         hostconfig = container.hostconfig
@@ -181,13 +212,20 @@ class Chef
         return hostconfig
       end
 
-      ## wrapper for tweaking the config if needed ##
+      ##
+      ## wrapper for tweaking the config if needed
+      ##
 
       def clean_config(container)
         return container.config 
       end
 
-      ## actions ##
+
+
+
+      ##
+      ## actions
+      ##
 
       def action_create
         ## create the new container
