+++ templates/default/client.rb.erb
@@ -4,9 +4,9 @@
 node_name                   "#{ENV['CHEF_NODE_NAME']}"
 chef_server_url             '<%= @chef_server_url %>'
 validation_client_name      '<%= @validation_client_name %>'
-validation_key              '<%= ::File.join(@chef_secure_path, 'validation.pem') %>'
-client_key                  '<%= ::File.join(@chef_secure_path, 'client_key.pem') %>'
-trusted_certs_dir           '<%= ::File.join(@chef_secure_path, 'trusted_certs') %>'
-encrypted_data_bag_secret   '<%= ::File.join(@chef_secure_path, 'encrypted_data_bag_secret') %>
+validation_key              '/etc/chef/secure/validation.pem'
+client_key                  '/etc/chef/secure/client_key.pem'
+trusted_certs_dir           '/etc/chef/secure/trusted_certs'
+encrypted_data_bag_secret   '/etc/chef/secure/encrypted_data_bag_secret'
 environment                 '<%= @chef_environment %>'
-ssl_verify_mode             :verify_none
+#ssl_verify_mode             :verify_none
